# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Money Management",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/utang": {
            "get": {
                "tags": [
                    "Utang"
                ],
                "summary": "Menampilkan semua data utang",
                "operationId": "getUtangList",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Utang"
                ],
                "summary": "Membuat data utang baru",
                "operationId": "storeUtang",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tanggal",
                                    "jumlah",
                                    "kreditur",
                                    "user_id"
                                ],
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "catatan": {
                                        "type": "string"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "format": "decimal"
                                    },
                                    "kreditur": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data utang berhasil dibuat"
                    }
                }
            }
        },
        "/utang/{id}": {
            "get": {
                "tags": [
                    "Utang"
                ],
                "summary": "Menampilkan detail utang berdasarkan ID",
                "operationId": "getUtangById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Utang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Utang"
                ],
                "summary": "Update data utang berdasarkan ID",
                "operationId": "updateUtang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Utang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "catatan": {
                                        "type": "string"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "format": "decimal"
                                    },
                                    "kreditur": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utang"
                ],
                "summary": "Hapus data utang berdasarkan ID",
                "operationId": "deleteUtang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Utang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                }
            }
        }
    },
...
}